+++ package.json
@@ -2,10 +2,10 @@
 	"name": "fagc-discord-bot",
 	"version": "0.0.1",
 	"description": "",
-	"main": "src/index.js",
+	"main": "dist/index.js",
 	"scripts": {
 		"lint": "npm run eslint",
-		"eslint": "eslint src/ --fix --ext .js",
+		"eslint": "eslint src/ --fix --ext .js,.ts",
 		"dev": "nodemon"
 	},
 	"dependencies": {
@@ -13,19 +13,20 @@
 		"@sentry/node": "^6.4.1",
 		"@sentry/tracing": "^6.4.1",
 		"@supercharge/strings": "^1.18.0",
+		"@types/dateformat": "^5.0.0",
 		"chalk": "^4.1.1",
 		"common-tags": "^1.8.0",
 		"dateformat": "^4.5.1",
-		"discord.js": "^12.5.1",
+		"discord.js": "^13.6.0",
 		"dotenv": "^8.2.0",
 		"envalid": "^7.2.1",
 		"fagc-api-wrapper": "^1.24.1",
 		"fs-extra": "^9.1.0",
 		"moment": "^2.29.1",
-		"mongoose": "^5.12.3",
 		"node-fetch": "^2.6.1",
 		"path": "^0.12.7",
 		"strict-uri-encode": "^2.0.0",
+		"typescript": "^4.5.4",
 		"validator": "^13.7.0"
 	},
 	"devDependencies": {
