--- src/utils/responses.js
@@ -1,20 +1,0 @@
-module.exports = {
-	// standard responses
-	error: {
-		TIME: "Times up! Please run the command again!",
-		CHANNEL: "Please specify a channel!",
-		USER: "Please mention a user!",
-		ERROR: "Something went wrong, please try again!",
-		TIMEARGS:
-			"Please enter a valid time! Available units: `s`, `m`, `h` or`d`",
-		EMPTY: "Please enter a message!",
-		NOID: "I can't find that id! Please try again!",
-		NOREASON: "No reason provided!",
-		STATUS: "Please enter either `on` or `off`",
-		ROLE: "Please specify a valid role!",
-		EMBEDLENGTH: "Your message must not exceed 1800 characters!",
-		SEARCH: "Please specify a search!",
-		NUMBER: "Please specify a number!",
-		ID: "Please specify an ID",
-	},
-}
